requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.12.0
requests==2.31.0
spacy==3.7.4
python-dateutil==2.8.2 